telegram/requirements/base.txt
Django==3.0.9
djangorestframework
django-redis
celery==4.4.5
requests==2.24.0